requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
tensorflow>=2.8.0
requests>=2.26.0
python-dotenv>=0.19.0
tqdm>=4.62.0
python-dateutil>=2.8.2
pytz>=2021.3
jupyter>=1.0.0
ipykernel>=6.0.0
black>=21.12b0
flake8>=4.0.0
mypy>=0.910
